refactor: move getPostSlugs and getPostBySlug to shared api-content file
