+++ src/app/(docs)/docs/changelog/[...slug]/page.jsx
@@ -7,7 +7,8 @@ import Link from 'components/shared/link';
 import VERCEL_URL from 'constants/base';
 import { CHANGELOG_DIR_PATH } from 'constants/content';
 import { CHANGELOG_BASE_PATH, CHANGELOG_SLUG_REGEX } from 'constants/docs';
-import { getAllChangelogs, getPostBySlug } from 'utils/api-docs';
+import { getPostBySlug } from 'utils/api-content';
+import { getAllChangelogs } from 'utils/api-docs';
 import getExcerpt from 'utils/get-excerpt';
 import getFormattedDate from 'utils/get-formatted-date';
 import getMetadata from 'utils/get-metadata';
