+++ src/app/(docs)/docs/changelog/rss.xml/route.js
@@ -3,7 +3,8 @@ import Rss from 'rss';
 
 import { CHANGELOG_DIR_PATH } from 'constants/content';
 import { CHANGELOG_BASE_PATH } from 'constants/docs';
-import { getAllChangelogs, getPostBySlug } from 'utils/api-docs';
+import { getPostBySlug } from 'utils/api-content';
+import { getAllChangelogs } from 'utils/api-docs';
 import getExcerpt from 'utils/get-excerpt';
 import getFormattedDate from 'utils/get-formatted-date';
