+++ src/app/[slug]/pages/template-page.jsx
@@ -6,7 +6,7 @@ import Container from 'components/shared/container';
 import Layout from 'components/shared/layout';
 import { TEMPLATE_PAGES_DIR_PATH } from 'constants/content';
 import { DEFAULT_IMAGE_PATH } from 'constants/seo-data';
-import { getPostBySlug } from 'utils/api-docs';
+import { getPostBySlug } from 'utils/api-content';
 import getMetadata from 'utils/get-metadata';
 import getTableOfContents from 'utils/get-table-of-contents';
