+++ src/app/faster/page.jsx
@@ -1,7 +1,7 @@
 import Post from 'components/pages/doc/post';
 import { FASTER_DIR_PATH } from 'constants/content';
 import SEO_DATA from 'constants/seo-data';
-import { getPostBySlug } from 'utils/api-docs';
+import { getPostBySlug } from 'utils/api-content';
 import getMetadata from 'utils/get-metadata';
 import getTableOfContents from 'utils/get-table-of-contents';
