+++ src/app/use-cases/[slug]/page.jsx
@@ -5,7 +5,7 @@ import Post from 'components/pages/template/post';
 import { USE_CASES_DIR_PATH } from 'constants/content';
 import LINKS from 'constants/links';
 import { DEFAULT_IMAGE_PATH } from 'constants/seo-data';
-import { getPostBySlug } from 'utils/api-docs';
+import { getPostBySlug } from 'utils/api-content';
 import getMetadata from 'utils/get-metadata';
 import getTableOfContents from 'utils/get-table-of-contents';
