+++ src/pages/api/release-notes.js
@@ -3,7 +3,7 @@ import fs from 'fs/promises';
 import matter from 'gray-matter';
 
 import { CHANGELOG_DIR_PATH } from 'constants/content';
-import { getPostSlugs } from 'utils/api-docs';
+import { getPostSlugs } from 'utils/api-content';
 import getExcerpt from 'utils/get-excerpt';
 
 // TODO: move this function to utils/api-docs
