+++ src/utils/api-docs.js
@@ -1,37 +1,10 @@
 const fs = require('fs');
 
-const { glob } = require('glob');
-const matter = require('gray-matter');
 const jsYaml = require('js-yaml');
 
 const { DOCS_DIR_PATH, CHANGELOG_DIR_PATH } = require('../constants/content');
 
-const getExcerpt = require('./get-excerpt');
-
-const getPostSlugs = async (pathname) => {
-  const files = await glob.sync(`${pathname}/**/*.md`, {
-    ignore: [
-      '**/RELEASE_NOTES_TEMPLATE.md',
-      '**/README.md',
-      '**/unused/**',
-      '**/shared-content/**',
-      '**/GUIDE_TEMPLATE.md',
-    ],
-  });
-  return files.map((file) => file.replace(pathname, '').replace('.md', ''));
-};
-
-const getPostBySlug = (slug, pathname) => {
-  try {
-    const source = fs.readFileSync(`${process.cwd()}/${pathname}/${slug}.md`, 'utf-8');
-    const { data, content } = matter(source);
-    const excerpt = getExcerpt(content, 200);
-
-    return { data, content, excerpt };
-  } catch (e) {
-    return null;
-  }
-};
+const { getPostSlugs, getPostBySlug } = require('./api-content');
 
 const getAllPosts = async () => {
   const slugs = await getPostSlugs(DOCS_DIR_PATH);
@@ -96,8 +69,6 @@ const getAllChangelogs = async () => {
 };
 
 module.exports = {
-  getPostSlugs,
-  getPostBySlug,
   getSidebar,
   getNavigationLinks,
   getAllChangelogs,
